+++ myems-api/core/point.py
@@ -224,7 +224,16 @@ class PointCollection:
             resp: Falcon response object
         """
         admin_control(req)
-
+        search_query = req.get_param('q', default=None)
+        if search_query is not None:
+            search_query = search_query.strip()
+        else:
+            search_query = ''
+        data_source_id = req.get_param('data_source_id', default=None)
+        if data_source_id is not None and data_source_id.isdigit() and int(data_source_id) > 0:
+            data_source_id = int(data_source_id)
+        else:
+            data_source_id = None
         # Redis cache key
         cache_key = 'point:list'
         cache_expire = 28800  # 8 hours in seconds (long-term cache)
@@ -276,7 +285,19 @@ class PointCollection:
                          "        high_limit, low_limit, higher_limit, lower_limit, ratio, offset_constant, "
                          "        is_trend, is_virtual, address, description, faults, definitions "
                          " FROM tbl_points ")
-                cursor.execute(query)
+                conditions = []
+                params = []
+                if data_source_id:
+                    conditions.append("data_source_id = %s")
+                    params.append(data_source_id)
+                if search_query:
+                    conditions.append("name LIKE %s")
+                    params.extend([f'%{search_query}%'])
+                if conditions:
+                    query += " WHERE " + " AND ".join(conditions)
+                query += " ORDER BY id "
+
+                cursor.execute(query,params)
                 rows = cursor.fetchall()
 
                 result = list()